requirements.txt
flask
flask-cors
numpy
openai
python-dotenv
pymongo